pyproject.toml
[project]
name = "granola-to-task"
version = "0.1.0"
description = "Extract action items from Granola meeting notes using AI"
readme = "README.md"
authors = [
    { name = "pnilan", email = "patrick.nilan@airbyte.io" }
]
requires-python = ">=3.14"
dependencies = [
    "airbyte-agent-granola>=0.1.3",
    "pydantic-ai-slim[anthropic]>=1.62.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
granola-to-task = "granola_to_task.main:main"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
